LATX: Move _top_in/_top_out to TranslationBlock
We try to refactor ETB to optimize some applications startup performance,
such as wpp. In this refactor we will move some member variables of ETB
to TranslationBlock, and remove functions and member variables that are
no longer used.

This patch is the first in ETB refactor series, includes the following
changes:
    1. Remove _top_increment from ETB
    2. Move _top_in/_top_out to TranslationBlock
    3. Remove unused functions:
       1) etb_set_top_in
       2) etb_get_top_out
       3) etb_set_top_out
       4) etb_get_top_increment
       5) lsenv_get_last_executed_tb_top_out

Later patches will show more details.

Signed-off-by: Hanlu Li <lihanlu@loongson.cn>