LATX: fix lock insn trigger segv scenario
lock insns acquire the lat_lock first, but in critical section, it may
trigger segv, cause the tb split and re-execute the lock insn.
There is no chance to unlock the lat_lock.

We modify the lock algorithm to check cpu_index, which keeps constant
during thread execution. If the lat_lock is held by the same cpu_index,
the lock is assumed to be already acquired.

Change-Id: I347066fe6a60da578889a96851715c65669e1bd4