LATX: fix unordered compare for fcom/fcomi etc..
In x86 SDM:
The FCOM instructions raise an invalid-arithmetic-operand exception
(#IA) when either or both of the operands is a NaN value or is in an
unsupported format. The FUCOM instructions perform the same operation
as the FCOM instructions, except that they do not generate an
invalid-arithmetic-operand exception for QNaNs.

We use SUN instead of CUN for these insns.

Change-Id: Ie9f0c3282bf99f791cade3d939b4deeaecdd7501