linux-user/mmap: Fix the SIGBUS for i386 guest on loongarch
This patch can avoid SIGBUS for i386 guest on loongarch (16K) host. We build
a simple test with -latomic on an i386 machine, such as helloworld, and run
it on loongarch (16K) host, the following error is thrown:
  $ ~/qemu-i386 ./hello-world
  qemu:handle_cpu_signal received signal outside vCPU context @ pc=0xfff658a07e

We found that the test failed because a SIGBUS is triggered in mmap_frag()
which is called by target_mmap(), and the SIGBUS is caused by memeset() write
the buffer that does not correspond to the file libatomic.so.1. As we know,
when mapping files into a memory area larger than the file, accesses to pages
beyond the file size will cause a SIGBUS. So if we emulating a target with a
smaller pagesize than the hosts, we need to add the PAGE_OVERFLOW prot to the
extra pages which is beyond the file size, and then remapping the extra pages
with MAP_ANONYMOUS when PAGE_OVERFLOW is detected in mmap_frag().

Some tests (helloworld, gcc atomic, wps) can run just fine on loongarch (16K)
after this patch.

Change-Id: I5c47d1e808aaf85c6ac7957ae7a0c6758b36c4d9
Signed-off-by: Hanlu Li <lihanlu@loongson.cn>