LATX: Add prefix_count in readPrefixes
If there is an instruction sequence with 2 prefixes, such as 0x66f3ab,
capstone can only record the second one for now. So add a prefix_count
to handle the situations with more than one prefix.

NOTE: We found the instruction sequence could be like this: 0xf366ab, capstone
can not handle this format becaus it assumes the prefix 0x66 is always at the
first byte, so we change the judging conditions about 0x66, to make sure that
we won't miss it.

This bug is detected by DBT test.

ISSUE: 20988
Change-Id: I5b0430bf984b05ac527f3ba49ad6e51dcafe0c76
Signed-off-by: Jinyang Shen <shenjinyang@loongson.cn>
(cherry picked from commit 9da2415b8d464dc0065e24f98902c05969ec7997)