LATX-SYS: sigint monitor native jmp glue 2 corner case
At the end of native jmp glue 2 we will check if there is an interrupt pending.

If the signal comes when vCPU is executing native jmp glue 2, the signal handler will
not know which TB's will be executed next, since it dynamically lookuped from jmp cache.
Even though we could get it from the native context's register after lookup, we still
do know before lookup finish.

So we simply check the interrupt before jump to next TB.

Still we need some instructions to read the interrupt flag. If the signal comes after
we read the interrupt flag and before we jump to next TB, we still can not response to
this interrupt since the we get the OLD interrupt flag value.

This corner case occurs rarely, for now we only print a warning for it.

Change-Id: I3a83c17dcb267e8ee1dc8c113ab8a4e3e04c0cca
Signed-off-by: NiuGenen <niugen@loongson.cn>