LATX-SYS: fix sigint signal handler could do nothing when vCPU is not executing TB
Add one flag @sigint_flag in CPUX86State to indicate whether vCPU thread is executing TB
> in context switch bt to native, right before jump to TB : set it to 0
> in context switch native to bt, right at the beginning  : set it to 1

When this flag is 0, it means vCPU is executing TB
When this flag is 1, it means vCPU is not executing TB

In signal handler, when it locates outside the code cache and the flag is 1, signal
handler could do nothing and return. In the previous commit, context switch bt to native
will always check interrupt before jump to TB. Therefore the interrupt will be responsed.
Also, in this situation, the @latxs_int_tb in CPUX86State is meanless. Unlink it will
not help.

Change-Id: Icd24c3358233f5428889ea37ecc6fd3dca094aac
Signed-off-by: NiuGenen <niugen@loongson.cn>